requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
aiohttp==3.9.5
websockets==12.0
python-dotenv==1.0.1
pydantic==2.7.0
jinja2==3.1.4
pytest==8.2.0
httpx==0.27.0